Revert "major visual update to projects sections and 'section' styling"
