--- static/css/main.css
@@ -1,333 +1,0 @@
-/**12/15/2017 2:55 AM should be just main.css**/
-
-
-/*Navigation bar*/
-nav, footer {
-    background-color: white;
-    position: fixed;
-    left: 0;
-    right: 0;
-    z-index: 9999;
-    text-align: center;
-    box-shadow: 0em .015em  .5em gray;
-}
-
-nav {
-    top:0;
-    padding-top: 15px;
-    padding-bottom: 15px;
-
-}
-
-footer {
-    bottom: 0;
-    padding-top: 10px;
-    padding-bottom: 10px;
-    
-}
-
-footer li {
-    padding: 0 20px;
-    align-self: center;
-}
-
-nav a.hamburgerMenu, footer a.hamburgerMenu {
-    display:none;
-}
-
-nav li, footer li{
-    display: inline-block;
-    margin: 0 0.5em;
-}
-
-nav a, footer a {    
-    text-decoration: none;
-    transition: box-shadow .5s, color .5s;
-    transition-delay: linear;
-    background-color: white;
-    color: rgb(72, 130, 187);
-    font-size: 1em;
-    
-}
-
-nav a:hover, footer a:hover {
-    color:rgb(100, 197, 230);
-    transition: 0.4s;
-
-}
-
-.card-details span {
-    vertical-align: bottom;
-
-}
-
-.card-details span button {
-    margin-right: 5px;
-    border: 1px solid lightgray;
-    border-radius: 3px;
-}
-
-
-
-
-
-
-/*BODY CONTENT*/
-
-body {
-    background-color: white;
-    font-family: 'Open Sans', sans-serif;
-    color: #3d3d3d;
-    font-size: 15px;
-    margin-bottom: 5em;
-    margin-left: 0;
-    margin-right: 0;
-
-}
-
-main {
-    padding-right: 0;
-    padding-left: 0;
-}
-
-
-section {
-    margin-top: 100px;
-    padding-left: 20%;
-    width: 60%;
-
-}
-
-.about-section {
-    border-left: 8px solid lightblue;
-}
-
-.about-section p {
-    line-height: 2;
-}
-
-.projects-section {
-    border-left: 8px solid skyblue;
-}
-
-
-mark{
-    background-color: lightblue;
-    padding: 0 3px;
-}
-
-.name {
-    font-size: 150%;
-}
-
-ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-header {
-    margin-top: 70px;
-    padding-left: 20%;
-    width: 60%;
-}
-
-h1 {
-    text-align: center;
-    font-family: 'Dosis', sans-serif;
-    margin-bottom: 0;
-    font-size: 50px;
-}
-
-h2 {
-    font-family: 'Dosis', sans-serif;
-    margin-top: 0;
-    border-bottom: 2px dotted black;
-}
-
-
-h3 {
-    margin-bottom: 0.25em;
-    font-size: 100%;
-    margin-top: 2em;
-}
-
-
-h3 + a:hover, p a:hover{
-    color:#ADD8E6;
-    font-style: oblique;
-}
-
-
-
-
-
-/*FOOTER BAR*/
-
-.resumeArrow {
-    width:1em;
-    vertical-align: top;
-    -webkit-transform: rotate(45deg);
-    -moz-transform: rotate(45deg);
-    -ms-transform: rotate(45deg);
-    -o-transform: rotate(45deg);
-    transform: rotate(45deg);
-}
-
-.resumeDownloadArrow {
-    width:1em;
-    vertical-align: top;
-    -webkit-transform: rotate(180deg);
-    -moz-transform: rotate(180deg);
-    -ms-transform: rotate(180deg);
-    -o-transform: rotate(180deg);
-    transform: rotate(180deg);
-    
-}
-
-.contactIcon {
-    width: 30px;
-    display: inline;
-    transform: scale(1);    
-    transition: transform 0.26s ease;
-}
-
-.contactIcon:hover {
-    transform: scale(1.4);
-}
-
-.contactText {
-    line-height: 2em;
-}
-
-a {
-    
-    text-decoration: none;
-}
-
-
-a:visited {
-    text-decoration: none;
-}
-
-
-
-
-
-/*CARD STYLING*/
-
-.card {
-    border-radius: 5px; /* 5px rounded corners */
-    border: 2px solid lightgray;
-    background: white;
-    height: 400px;
-    width: 230px;
-    margin-top: 10px;
-    margin-left: 0%;
-    margin-right: 0%;
-    margin-bottom: 5px;
-    overflow: hidden;
-}
-
-.card:hover {
-    box-shadow: 0 4px 8px 5px rgba(0,0,0,0.2);
-    transition: 0.5s;
-}
-
-.card-details {
-    font-size: 12px;
-    padding: 0 10px;
-    overflow: hidden;
-}
-
-.card-details p {
-    line-height: 20px;
-}
-
-.card img {
-    width: 100%;
-    height: 230px;
-    border-bottom: 2px solid lightgray;
-}
-
-.flex-container {
-    display: flex;
-    margin-bottom: 300px;
-    flex-flow: row wrap;
-    justify-content: space-between;
-}
-
- .card-details h4, p {
-    margin: 2px 0;
-}
-
-
-
-
-
-/*--------Breakpoints and media queries for responsive design--------*/
-
-/*----Smallest screens----*/
-@media (max-width:760px){
-
-    nav {
-        padding-top: .7em;
-        padding-bottom: .7em;
-    }
-    
-    nav li {
-        margin: 0 .15em;
-    }
-
-    h1 {
-        margin-top: 2em;
-    }
-    
-    h2 {
-        margin-bottom: .25em;
-    }
-    
-    h3 {
-        margin-top: .25em;
-    }
-    
-    .name {
-        font-size: 125%;
-    }
-    
-    body {
-        font-size: 95%;
-    }
-    
-}
-
-
-
-
-/*----Medium sized screens: laptops, tablets----*/
-@media (min-width:760px) and (max-width:1480px){
-    main{
-        padding-right: 0;
-        padding-left: 0;
-    }      
-    
-}
-
-
-
-
-
-/*----Larger desktop screens----*/
-@media (min-width:1480px){
-    body{
-    }
-    
-    nav li {
-        margin: 0 1em;
-    }
-    
-    h1 {
-        margin-top: 15%;
-    }
-
-}
